+++ demo-main/build_ex.sh
@@ -11,6 +11,7 @@ CANV="$EXDIR/canvas"
 SRCS="../components/fxtk/fxtk.c ../components/fxtk/fxtk_draw.c \
       ../components/fxtk/fxtk_widgets.c ../components/fxtk/fxtk_font.c \
       ../components/fxtk/fxtk_effects.c ../components/fxtk/fxtk_extra.c \
+      ../components/fxtk/fxtk_fs.c \
       fxtk_sdl_driver.c fxtk_image_sdl.c main_linux.c"
 LIBS="-lSDL2 -lSDL2_ttf -lSDL2_image -lm -pthread -lX11"